+++ src/components/bluepad32/controller/uni_controller.c
@@ -16,19 +16,26 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ****************************************************************************/
 
-#include "uni_keyboard.h"
-#include "hid_usage.h"
+#include "controller/uni_controller.h"
 #include "uni_log.h"
 
-void uni_keyboard_dump(const uni_keyboard_t* kb) {
-    // Don't add "\n"
-    logi("modifiers=%#x, pressed keys=[", kb->modifiers);
-    for (int i = 0; i < UNI_KEYBOARD_PRESSED_KEYS_MAX; i++) {
-        if (kb->pressed_keys[i] == HID_USAGE_KB_NONE || kb->pressed_keys[i] == HID_USAGE_KB_ERROR_ROLL_OVER)
+void uni_controller_dump(const uni_controller_t* ctl) {
+    switch (ctl->klass) {
+        case UNI_CONTROLLER_CLASS_BALANCE_BOARD:
+            uni_balance_board_dump(&ctl->balance_board);
             break;
-        if (i > 0)
-            logi(", ");
-        logi("%#x", kb->pressed_keys[i]);
+        case UNI_CONTROLLER_CLASS_GAMEPAD:
+            uni_gamepad_dump(&ctl->gamepad);
+            break;
+        case UNI_CONTROLLER_CLASS_MOUSE:
+            uni_mouse_dump(&ctl->mouse);
+            break;
+        case UNI_CONTROLLER_CLASS_KEYBOARD:
+            uni_keyboard_dump(&ctl->keyboard);
+            break;
+        default:
+            logi("uni_controller_dump: Unsupported controller class: %d\n", ctl->klass);
+            break;
     }
-    logi("]");
+    logi(", battery=%d\n", ctl->battery);
 }