+++ src/components/bluepad32/parser/uni_hid_parser_atari.c
@@ -21,9 +21,9 @@ limitations under the License.
 
 #include "parser/uni_hid_parser_atari.h"
 
+#include "controller/uni_controller.h"
 #include "hid_usage.h"
 #include "uni_common.h"
-#include "uni_controller.h"
 #include "uni_hid_device.h"
 #include "uni_log.h"