+++ src/components/bluepad32/parser/uni_hid_parser_wii.c
@@ -33,9 +33,9 @@ limitations under the License.
 
 #include "parser/uni_hid_parser_wii.h"
 
+#include "controller/uni_controller.h"
 #include "hid_usage.h"
 #include "uni_common.h"
-#include "uni_controller.h"
 #include "uni_hid_device.h"
 #include "uni_log.h"