+++ src/components/bluepad32/parser/uni_hid_parser_xboxone.c
@@ -24,8 +24,8 @@ limitations under the License.
 
 #include "parser/uni_hid_parser_xboxone.h"
 
+#include "controller/uni_controller.h"
 #include "hid_usage.h"
-#include "uni_controller.h"
 #include "uni_hid_device.h"
 #include "uni_log.h"