+++ src/components/bluepad32/platform/uni_platform_mightymiggy.c
@@ -29,9 +29,9 @@ limitations under the License.
 #include "sdkconfig.h"
 
 #include "bt/uni_bt.h"
+#include "controller/uni_controller.h"
 #include "uni_common.h"
 #include "uni_config.h"
-#include "uni_gamepad.h"
 #include "uni_hid_device.h"
 #include "uni_hid_device_vendors.h"
 #include "uni_joystick.h"